+++ Classes/Creators/FieldControls.php
@@ -79,7 +79,7 @@ class FieldControls{
 
 
 		$in_standard = array( 'text', 'textarea', 'email', 'checkbox', 'number', 'checkboxes', 'radio', 'select' );
-		$in_adv = array( 'file', 'wysiwyg', 'date', 'hidden', 'address' );
+		$in_adv = array( 'file', 'wysiwyg', 'date', 'password', 'hidden', 'address' );
 
 		$in_des = array( 'html', 'break' );